career-connect-ai/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==7.4.4
redis==5.0.0
aiohttp==3.9.1